Moved make_mujoco_env to source code
Also minor adjustment in handling envpool import
